Enforcing float32 precision on ConfigSpace floats

+++ file_collator.py
@@ -49,11 +49,11 @@ def update_table_with_new_entry(
     seed = res['info']['seed']
     key_nest = []
     for k, v in config.items():
-        key_nest.append(v)
+        key_nest.append(np.float32(v))
         if glom.glom(main_data, glom.Path(*key_nest), default=None) is None:
             glom.assign(main_data, glom.Path(*key_nest), dict())
     for k, v in fidelity.items():
-        key_nest.append(v)
+        key_nest.append(np.float32(v))
         if glom.glom(main_data, glom.Path(*key_nest), default=None) is None:
             glom.assign(main_data, glom.Path(*key_nest), dict())
     key_nest.append(seed)
